fix(ci): install Vulkan SDK on Windows; fix ubuntu link + flaky profiler test
- setup-build-deps（Windows）：安装 LunarG Vulkan SDK 1.4.304.1
  （actions/cache 缓存 C:\VulkanSDK，按 vulkan.h 实际位置探测后导出
  VULKAN_SDK），使 CI 完整编译 Vulkan 后端而非跳过。
- dag_serializer: get<int64_t>() 在 Linux/gcc 下是 get<long>，实例化
  无定义主模板导致链接失败（macOS 的 int64_t 是 long long 恰有特化），
  改用 get<long long>()。
- test_profiler: TaskStatsAggregation 的顺序断言（A 时长 > B）在 CI
  负载下 flaky，改为按各自 sleep 的下限断言。
- bump submodules/scripting -> d56acd8（quickjs MSVC C2099 修复）。

+++ src/dag_serializer.cpp
@@ -103,7 +103,9 @@ TaskParams parse_params_with_specs(const nlohmann::json& params_json,
         } else {
             // 未声明的 key：按 JSON 字面量类型回退
             if (value.is_number_integer())
-                params.set_int(key, static_cast<int>(value.get<int64_t>()));
+                // 用 get<long long> 而非 get<int64_t>：后者在 Linux/gcc 下是 long，
+                // 会实例化无定义的主模板（macOS 的 int64_t 是 long long 恰好有特化）。
+                params.set_int(key, static_cast<int>(value.get<long long>()));
             else if (value.is_number_float())
                 params.set_float(key, static_cast<float>(value.get<double>()));
             else if (value.is_string())
